Split collections into helper classes
