+++ speckle_connector/src/convertors/to_speckle.rb
@@ -12,7 +12,7 @@ require_relative '../speckle_objects/other/rendering_options'
 require_relative '../speckle_objects/built_elements/view3d'
 require_relative '../speckle_objects/built_elements/revit/direct_shape'
 require_relative '../speckle_objects/relations/layers'
-require_relative '../speckle_objects/speckle/core/models/collection'
+require_relative '../speckle_objects/speckle/core/models/model_collection'
 require_relative '../constants/path_constants'
 require_relative '../sketchup_model/reader/speckle_entities_reader'
 require_relative '../sketchup_model/query/entity'
@@ -21,7 +21,7 @@ module SpeckleConnector
   module Converters
     # Converts sketchup entities to speckle objects.
     class ToSpeckle < Converter
-      COLLECTION = SpeckleObjects::Speckle::Core::Models::Collection
+      MODEL_COLLECTION = SpeckleObjects::Speckle::Core::Models::ModelCollection
       DIRECT_SHAPE = SpeckleObjects::BuiltElements::Revit::DirectShape
       SPECKLE_ENTITIES_READER = SketchupModel::Reader::SpeckleEntitiesReader
       VIEW3D = SpeckleObjects::BuiltElements::View3d
@@ -30,8 +30,8 @@ module SpeckleConnector
       # @return [Hash{Symbol=>Array}] layers -which only have objects- to hold it's objects under the base object.
       def convert_selection_to_base(preferences)
         convert = method(:convert)
-        new_speckle_state, model_collection = COLLECTION.layers(sketchup_model, speckle_state, @units, preferences,
-                                                                &convert)
+        new_speckle_state, model_collection = MODEL_COLLECTION.from_sketchup_model(sketchup_model, speckle_state,
+                                                                                   @units, preferences, &convert)
 
         # send only layers that have any object
         model_collection[:layers_relation] = SpeckleObjects::Relations::Layers.from_model(sketchup_model)
