+++ speckle_connector/src/speckle_objects/relations/layer.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../base'
+require_relative '../other/color'
 
 module SpeckleConnector
   module SpeckleObjects
